requirements.txt
langchain
openai
faiss-cpu
python-dotenv
PyMuPDF